Move design stats to a separate file
